pyproject.toml
[project]
name = "py-instrument"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.3",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.35.0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "opentelemetry-instrumentation-asyncpg>=0.56b0",
    "opentelemetry-instrumentation-fastapi>=0.56b0",
    "opentelemetry-instrumentation-httpx>=0.56b0",
    "opentelemetry-instrumentation-logging>=0.56b0",
    "opentelemetry-instrumentation-redis>=0.56b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.56b0",
    "opentelemetry-sdk>=1.35.0",
    "python-json-logger>=3.3.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
